- Set up basic telemetry test - Applied a few miscellaneous fixes to get telemetry to work, the most problematic of which is the DataBuffers in the TelemetryEndpoint class (they probably need to be Lists).
