+++ gradle.properties
@@ -19,4 +19,7 @@ use_access_transformer = false
 
 # Coremod Arguments
 include_mod = true
-coremod_plugin_class_name = 
+coremod_plugin_class_name =
+
+# JUnit testing toggle
+enableJUnit = true
