+++ src/main/java/com/starl0stgaming/gregicalitystarbound/api/telemetry/network/packet/data/TelemetryPacketPayload.java
@@ -10,6 +10,9 @@ public class TelemetryPacketPayload {
     public TelemetryPacketPayload() {
         this.nbtTagCompound = new NBTTagCompound();
     }
+    public TelemetryPacketPayload(NBTTagCompound nbtTagCompound) {
+        this.nbtTagCompound = nbtTagCompound;
+    }
 
     public NBTTagCompound getPayload() {
         return nbtTagCompound;
